delete forkRepository and move shared functions to abstract version control service
